Extract engine helper funcs, add test for one
